// GM_ClientSide/src/app/core/generated.ts
export interface NotificationModel {
    title?: string | undefined;
    message?: string | undefined;
    url?: string | undefined;
}